6_kafka/requirements.txt
kafka-python
confluent_kafka
requests
avro
faust
fastavro